project end

+++ src/routes/links.js
@@ -8,14 +8,14 @@ router.get('/add', isLoggedIn, (req, res) => {
 })
 
 router.get('/', isLoggedIn, async (req, res) => {
-  const links = await db.query('SELECT * FROM links')
+  const links = await db.query('SELECT * FROM links WHERE user_id = ?', [req.user.id])
   res.render('links/list', { links })
 })
 
 router.post('/add', isLoggedIn, async (req, res) => {
   const { title, url, description } = req.body
   const newLink = {
-    title, url, description
+    title, url, description, user_id: req.user.id
   }
   await db.query('INSERT INTO links set ?', [newLink])
   req.flash('success', 'Link saved successfully')
@@ -24,14 +24,14 @@ router.post('/add', isLoggedIn, async (req, res) => {
 
 router.get('/delete/:id', isLoggedIn, async (req, res) => {
   const { id } = req.params
-  await db.query('DELETE FROM links WHERE id = ?', [id])
+  await db.query('DELETE FROM links WHERE id = ? and user_id = ?', [id, req.user.id])
   req.flash('success', 'Links removed successfully')
   res.redirect('/links')
 })
 
 router.get('/edit/:id', isLoggedIn, async (req, res) => {
   const { id } = req.params
-  const link = await db.query('SELECT * FROM links WHERE id = ?', [id])
+  const link = await db.query('SELECT * FROM links WHERE id = ? and user_id = ?', [id, req.user.id])
   res.render('links/edit', { link: link[0] })
 })
